Add powerful setting analytical model

+++ nms/cpu_nms.pyx
@@ -17,7 +17,7 @@ cdef inline np.float32_t max(np.float32_t a, np.float32_t b):
 cdef inline np.float32_t min(np.float32_t a, np.float32_t b):
     return a if a <= b else b
 
-def cpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh):
+def cpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float32 thresh):
     cdef np.ndarray[np.float32_t, ndim=1] x1 = dets[:, 0]
     cdef np.ndarray[np.float32_t, ndim=1] y1 = dets[:, 1]
     cdef np.ndarray[np.float32_t, ndim=1] x2 = dets[:, 2]
